Update MenuItem/ListItem link components
